requirements.txt
beautifulsoup4=4.12.2
requests=2.31.0
six=1.16.0
python-dotenv=1.0.0